Test E2E test run lock for volume provisioner
